core: sel1 spmc: boot abi update
Updates the boot ABI for S-EL1 SPMC to align better with other SPMCs,
like Hafnium, but also with the non-FF-A configuration.

Register usage:
X0 - TOS FW config address, if not NULL
X2 - System DTB, if not NULL

Note that this is an incompatible change and requires corresponding
change in TF-A ("feat(qemu): update abi between spmd and spmc").

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>